% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_processing.R
\name{apply_hanning_window}
\alias{apply_hanning_window}
\title{applies a hanning window}
\usage{
apply_hanning_window(rec)
}
\arguments{
\item{equalized_signal}{the signal should be already in matrix form}
}
\value{
it modifies the signal in place
}
\description{
applies a hanning window
}
