% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_files.R
\name{apply_indexes}
\alias{apply_indexes}
\title{applies an interval to a set of signals}
\usage{
apply_indexes(signals, indexes)
}
\arguments{
\item{signals}{a signals list object}

\item{indexes}{an interval index object}
}
\value{
it appends signal interval to the set
}
\description{
applies an interval to a set of signals
}
\examples{
files<-list(
system.file("extdata", "Grabacion_2_andrea.wav", package = "soundlocalization"),
system.file("extdata", "Grabacion_1_joaco.wav", package = "soundlocalization"),
system.file("extdata", "Grabacion_1_cande.wav", package = "soundlocalization"),
system.file("extdata", "Grabacion_3_luciano.wav", package = "soundlocalization")
)
waveList=loadwavfiles(files)
indexes=indexes_for_same_length(waveList)
aligned=apply_indexes(waveList,indexes)

}
