% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_files.R
\name{calc_lag_ccf}
\alias{calc_lag_ccf}
\title{calculates optimal lags for each microphone}
\usage{
calc_lag_ccf(max_lag)
}
\arguments{
\item{max_ccf}{list containing the paired lags}
}
\value{
it appends the optimal lags for each microphone pair
}
\description{
calculates optimal lags for each microphone
}
