% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_files.R
\name{calculate_ffts}
\alias{calculate_ffts}
\title{Title obtains the fourier transform of all the signals at once}
\usage{
calculate_ffts(windowed_signal)
}
\arguments{
\item{windowed_signal}{the hanning window should be already applied}
}
\value{
inserts a field with the fft
}
\description{
Title obtains the fourier transform of all the signals at once
}
