% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_files.R
\name{calculate_list_of_cross_correlations}
\alias{calculate_list_of_cross_correlations}
\title{Title calculates all the cross correlation pairs}
\usage{
calculate_list_of_cross_correlations(ffted_signal)
}
\arguments{
\item{ffted_signal}{the fft should be already calculated}
}
\value{
inserts the cross correlation function
}
\description{
Title calculates all the cross correlation pairs
}
