% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_files.R
\name{index_to_lag}
\alias{index_to_lag}
\title{convert an index of a ccf to the corresponding lag}
\usage{
index_to_lag(i, nsamples)
}
\arguments{
\item{i}{index of a ccf}

\item{nsmaples}{number of samples of the ccf}
}
\value{
the lag that correspond to the index i
}
\description{
convert an index of a ccf to the corresponding lag
}
