% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_files.R
\name{indexes_for_same_length}
\alias{indexes_for_same_length}
\title{calculates the indexes so all the waves are of the same size}
\usage{
indexes_for_same_length(signals)
}
\arguments{
\item{list_of_waves}{a list of class Wave objects}
}
\value{
a list with the start end indexes vectors
}
\description{
calculates the indexes so all the waves are of the same size
}
\examples{
files<-list(
system.file("extdata", "Grabacion_2_andrea.wav", package = "soundlocalization"),
system.file("extdata", "Grabacion_1_joaco.wav", package = "soundlocalization"),
system.file("extdata", "Grabacion_1_cande.wav", package = "soundlocalization"),
system.file("extdata", "Grabacion_3_luciano.wav", package = "soundlocalization")
)
waveList=loadwavfiles(files)
indexes=indexes_for_same_length(waveList)


}
