% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadwavfiles.R
\name{loadwavfiles}
\alias{loadwavfiles}
\title{Load several wav files at once}
\usage{
loadwavfiles(fileList)
}
\arguments{
\item{list}{of files to be loaded}
}
\value{
a list of wave objects
}
\description{
Load several wav files at once
}
\examples{
files<-list("~/papers/ranas/recordings/venezuela 0/Grabacion_3_luciano.wav")
d<-loadwavfiles(files)


}
