% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_files.R
\name{max_lag_ccf}
\alias{max_lag_ccf}
\title{finds the lag that correspond to the maximal cross correlation}
\usage{
max_lag_ccf(signal)
}
\arguments{
\item{ccf}{a list that contains the pairs of cross correlations}
}
\value{
inserts that optimal lags of each pair of microphones
}
\description{
finds the lag that correspond to the maximal cross correlation
}
