% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_files.R
\name{optimize_delays_from_ccf}
\alias{optimize_delays_from_ccf}
\title{Calculates the optimal lag combination for a set of ccfs pairs}
\usage{
optimize_delays_from_ccf(maxs_ccf)
}
\arguments{
\item{ccf_list}{should include double indexed list of cross correlation functions}
}
\value{
it inserts the global and local optimal lags
}
\description{
Calculates the optimal lag combination for a set of ccfs pairs
}
